requirements.txt
numpy==1.23
scipy==1.11.1
numba==0.57.1
pyyaml
librosa
tqdm
webrtcvad
torchmetrics
einops
thop
pynvrtc==9.2
pesq
pystoi
matplotlib
tensorboard==2.13
